sidecar/llama-zip/requirements.txt
grpcio==1.72.0
grpcio-tools==1.72.0
protobuf==5.29.4
llama-zip==0.2.1
llama-cpp-python==0.3.9
huggingface-hub==0.30.2
